Remove tpmsim tag from example cmd and rely on CGO_ENABLED flag instead
